backend/requirements.txt
fastapi
uvicorn[standard]
ccxt
pandas
python-dotenv
duckdb
pydantic
optuna
passlib[argon2]
python-jose[cryptography]
python-multipart
cryptography
